api/presentation/requirements.txt
litestar[standard]>=2.0,<3.0
uvicorn>=0.30,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
aiosqlite>=0.20,<1.0
openpyxl>=3.1,<4.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
